refactor: remove outdated code
