--- src/rigidTransform/createRigidTransformModel.ts
@@ -1,29 +1,0 @@
-import { ModelFunction } from '..';
-
-import { Point } from '../toMigrate/forImageJs/getMatrixFromPoints';
-
-/**
- * Generate a function that applies the given transformation parameters to a point.
- *
- * @param transform - Transformation to apply.
- * @returns Transformed point.
- */
-export function createRigidTransformModel(
-  transform: number[],
-): ModelFunction<Point> {
-  return (point: Point) => {
-    const angle = (transform[0] * Math.PI) / 180;
-    const xTranslation = transform[1];
-    const yTranslation = transform[2];
-    const column =
-      Math.cos(angle) * point.column -
-      Math.sin(angle) * point.row +
-      xTranslation;
-    const row =
-      Math.sin(angle) * point.column +
-      Math.cos(angle) * point.row +
-      yTranslation;
-
-    return { column, row };
-  };
-}
